knapsack algorithm
A thief wants to rob a store. He is carrying a bag of capacity W. The store has ‘n’ items. Its weight is given by an array and its value by another array. He can either include an item in its knapsack or exclude it but can’t partially have it as a fraction. We need to find the maximum value of items that the thief can steal.
use of knapsack algorithm.