PongGame.py
Most of us have heard about the famous pong game. Many of us love playing it. 
Before starting with the coding part we first need to install the turtle module. The turtle module is a Python library that enables users to create pictures and shapes by providing them with a virtual canvas.